Create riamusource
I'd say I'm about 30% done with this game. It's basically a story about your character - mage, warrior or archer - and the different adventures you can go on in Riamu. I'm about halfway done with the mage story.
FYI I came up with the name because it's my name in Japanese.